@@ -1,7 +1,7 @@
 name: android_semantics_testing
 description: Integration testing library for Android semantics
 environment:
-  sdk: '>=2.9.0 <3.0.0'
+  sdk: '>=2.17.0-0 <3.0.0'
 
 dependencies:
   flutter:
