@@ -10,7 +10,7 @@ export 'popup_constants.dart';
 
 /// A page with a popup menu, a dropdown menu, and a modal alert.
 class PopupControlsPage extends StatefulWidget {
-  const PopupControlsPage({Key key}) : super(key: key);
+  const PopupControlsPage({super.key});
 
   @override
   State<StatefulWidget> createState() => _PopupControlsPageState();
@@ -60,9 +60,9 @@ class _PopupControlsPageState extends State<PopupControlsPage> {
                     child: Text(item),
                   );
                 }).toList(),
-                onChanged: (String value) {
+                onChanged: (String? value) {
                   setState(() {
-                    dropdownValue = value;
+                    dropdownValue = value!;
                   });
                 },
               ),
